requirements.txt
# Core ML / NLP
torch>=2.1.0
transformers>=4.40.0
sentence-transformers>=2.6.0
peft>=0.10.0
trl>=0.8.0
bitsandbytes>=0.43.0
datasets>=2.18.0
accelerate>=0.29.0

# Vector search
faiss-cpu>=1.8.0

# Config & utilities
pyyaml>=6.0
numpy>=1.26.0
tqdm>=4.66.0

# Optional: OpenAI-based claim extraction
openai>=1.14.0
